[CS-1722] - Split hook from screen to test it
