+++ cardstack/src/screens/index.ts
@@ -4,6 +4,6 @@ export { default as MerchantScreen } from './MerchantScreen';
 export { default as TransactionConfirmation } from './TransactionConfirmation';
 export { default as CopyAddressSheet } from './CopyAddressSheet';
 export { default as PrepaidCardModal } from './PrepaidCardModal';
-export { default as SendSheetDepot } from './SendSheetDepot';
+export { default as SendSheetDepot } from './SendSheetDepot/SendSheetDepot';
 export { default as ShowQRCodeModal } from './ShowQRCodeModal';
 export { default as PayMerchantUniversalLink } from './PayMerchantUniversalLink';
